[feat] 홈, 온보딩 패키지 정리
