+++ app/src/main/java/com/example/soullive/home/HomeAiModelAdapter.kt
@@ -1,4 +1,4 @@
-package com.example.soullive
+package com.example.soullive.home
 
 import android.content.Intent
 import android.view.LayoutInflater
@@ -8,16 +8,17 @@ import android.widget.ImageView
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
+import com.example.soullive.R
 
-class HomeKeywordModelAdapter(private var models: List<Model>)
-    : RecyclerView.Adapter<HomeKeywordModelAdapter.HomKeywordModelViewHolder>() {
+class HomeAiModelAdapter(private var models: List<Model>)
+    : RecyclerView.Adapter<HomeAiModelAdapter.HomeAiModelViewHolder>() {
 
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): HomKeywordModelViewHolder {
-        val view = LayoutInflater.from(parent.context).inflate(R.layout.recycler_view_home_keyword_model, parent, false)
-        return HomKeywordModelViewHolder(view)
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): HomeAiModelViewHolder {
+        val view = LayoutInflater.from(parent.context).inflate(R.layout.recycler_view_home_ai_model, parent, false)
+        return HomeAiModelViewHolder(view)
     }
 
-    override fun onBindViewHolder(holder: HomKeywordModelViewHolder, position: Int) {
+    override fun onBindViewHolder(holder: HomeAiModelViewHolder, position: Int) {
         val model = models[position]
         holder.bind(model)
     }
@@ -32,11 +33,10 @@ class HomeKeywordModelAdapter(private var models: List<Model>)
         notifyDataSetChanged()
     }
 
-    inner class HomKeywordModelViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
+    inner class HomeAiModelViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
         private val modelImage: ImageView = itemView.findViewById(R.id.model_img)
         private val modelName: TextView = itemView.findViewById(R.id.model_name)
-        private val companyName: TextView = itemView.findViewById(R.id.company_name)
-        private val title: TextView = itemView.findViewById(R.id.title)
+        private val modelType: TextView = itemView.findViewById(R.id.model_type)
 
         init {
             itemView.setOnClickListener {
@@ -50,6 +50,7 @@ class HomeKeywordModelAdapter(private var models: List<Model>)
 
         fun bind(modelList: Model) {
             modelName.text = modelList.name
+            modelType.text = modelList.job
 
             Glide.with(itemView.context)
                 .load(modelList.image)
@@ -57,6 +58,5 @@ class HomeKeywordModelAdapter(private var models: List<Model>)
                 .error(R.drawable.img_model_sample) // 에러 이미지 리소스
                 .into(modelImage)
         }
-
     }
 }
