+++ app/src/main/java/com/example/soullive/onboarding/BottomCompanyTypeFragment.kt
@@ -1,11 +1,11 @@
-package com.example.soullive
+package com.example.soullive.onboarding
 
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.TextView
+import com.example.soullive.R
 import com.google.android.material.bottomsheet.BottomSheetDialog
 import com.google.android.material.bottomsheet.BottomSheetDialogFragment
