+++ app/src/main/res/layout/activity_onboarding_done.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/login_bg"
-    tools:context=".OnboardingDoneActivity">
+    tools:context=".onboarding.OnboardingDoneActivity">
 
     <ImageView
         android:layout_width="700dp"
