+++ app/src/main/res/layout/fragment_home.xml
@@ -7,7 +7,7 @@
     android:background="@color/bg"
     android:orientation="vertical"
     android:padding="15dp"
-    tools:context=".HomeFragment">
+    tools:context=".home.HomeFragment">
 
     <ScrollView
         android:layout_width="match_parent"
