+++ app/src/main/res/layout/fragment_view_pager1.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:background="#00000000"
     android:orientation="vertical"
-    tools:context=".ViewPager1Fragment">
+    tools:context=".home.ViewPager1Fragment">
 
     <LinearLayout
         android:layout_width="match_parent"
