+++ app/src/main/res/layout/fragment_view_pager2.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#00000000"
-    tools:context=".ViewPager2Fragment">
+    tools:context=".home.ViewPager2Fragment">
 
     <LinearLayout
         android:layout_width="match_parent"
